pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bazinga-indeed"
version = "6.1.0"
description = "BAZINGA - The first AI you actually own. Free, private, works offline. Multi-AI consensus through φ-coherence. TrD + TD = 1."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Abhishek Srivastava", email = "bits.abhi@gmail.com"}
]
keywords = [
    "ai", "distributed", "consciousness", "phi", "golden-ratio",
    "rag", "embeddings", "llm", "groq", "vector-database",
    "federated-learning", "lora", "differential-privacy", "p2p",
    "quantum", "lambda-g", "tensor", "vac", "emergence",
    "blockchain", "proof-of-boundary", "darmiyan", "consensus", "triadic",
    "inter-ai", "multi-ai", "consensus", "claude", "gemini", "ollama",
    "carm", "catastrophic-forgetting", "prime-lattice", "resonant-memory",
    "trust-dimension", "trd", "consciousness-scaling", "phi-coherence"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0,<3.0.0",
    "chromadb>=0.4.0,<1.0.0",
    "sentence-transformers>=2.2.0,<4.0.0",
    "httpx>=0.24.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
    "pyzmq>=25.0.0,<27.0.0",
    "nest-asyncio>=1.5.0,<2.0.0",
    "textual>=0.50.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]
local = [
    "llama-cpp-python>=0.2.0",
    "huggingface_hub>=0.20.0",
]
mcp = [
    "mcp[cli]>=1.0.0",
]
federated = [
    "torch>=2.0.0",
]
full = [
    "torch>=2.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "llama-cpp-python>=0.2.0",
    "huggingface_hub>=0.20.0",
    "mcp[cli]>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/0x-auth/bazinga-indeed"
Documentation = "https://github.com/0x-auth/bazinga-indeed#readme"
Repository = "https://github.com/0x-auth/bazinga-indeed"
Issues = "https://github.com/0x-auth/bazinga-indeed/issues"
Donate = "https://github.com/0x-auth/bazinga-indeed/blob/main/DONATE.md"
"Donate (India/UPI)" = "https://razorpay.me/@bitsabhi"
"Donate (Crypto)" = "https://etherscan.io/address/0x720ceF54bED86C570837a9a9C69F1Beac8ab8C08"

[project.scripts]
bazinga = "bazinga:main_sync"
bazinga-mcp = "bazinga.mcp_server:run_server"

[tool.setuptools.packages.find]
where = ["."]
include = ["bazinga*", "src*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]
